food_beverage_service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
